+++ src/zandronum/src/gl/utility/gl_convert.h
@@ -3,7 +3,7 @@
 #define __GLC_CONVERT
 
 #include "m_fixed.h"
-#define ANGLE_TO_FLOAT(ang) ((float)((ang) * 180. / ANGLE_180))
+#define ANGLE_TO_FLOAT(ang) ((float)(double(ang) * 180. / ANGLE_180))
 #define FLOAT_TO_ANGLE(ang) xs_RoundToUInt((ang) / 180. * ANGLE_180)
 
 #endif
